refactor: move publish to CLI command, auto-version deploy
- Move publish-artifact.ts to src/commands/publish.ts
- Deploy now reads version from package.json (use --dev for dev releases)
- Move @aws-sdk/client-s3 to dependencies
